Decompose map model and render planning files #297
