+++ Sources/Features/Map/MapFeatureModel.swift
@@ -327,959 +327,3 @@ final class MapFeatureModel {
         )
     }
 }
-
-struct MapLayerScene {
-    let canvasState: MapCanvasState
-    let legendState: MapLegendState
-    let warningLegendItems: [WarningLegendItem]
-
-    static func placeholder(
-        for layer: MapLayer,
-        initialCenterCoordinate: CLLocationCoordinate2D? = nil
-    ) -> MapLayerScene {
-        MapLayerScene(
-            canvasState: MapCanvasState(
-                overlays: [],
-                overlayRevision: 0,
-                initialCenterCoordinate: initialCenterCoordinate
-            ),
-            legendState: .loading(for: layer),
-            warningLegendItems: []
-        )
-    }
-
-    func withInitialCenterCoordinateIfNeeded(_ coordinate: CLLocationCoordinate2D) -> MapLayerScene {
-        guard canvasState.initialCenterCoordinate == nil else { return self }
-        return MapLayerScene(
-            canvasState: canvasState.withInitialCenterCoordinate(coordinate),
-            legendState: legendState,
-            warningLegendItems: warningLegendItems
-        )
-    }
-
-    func withPresentationState(_ presentationState: MapLegendPresentationState) -> MapLayerScene {
-        MapLayerScene(
-            canvasState: canvasState,
-            legendState: legendState.withPresentationState(presentationState),
-            warningLegendItems: warningLegendItems
-        )
-    }
-}
-
-struct MapCanvasState {
-    let overlays: [MapOverlayEntry]
-    let overlayRevision: Int
-    let initialCenterCoordinate: CLLocationCoordinate2D?
-
-    func withInitialCenterCoordinate(_ coordinate: CLLocationCoordinate2D) -> MapCanvasState {
-        MapCanvasState(
-            overlays: overlays,
-            overlayRevision: overlayRevision,
-            initialCenterCoordinate: coordinate
-        )
-    }
-}
-
-struct MapLegendState: Sendable {
-    let presentationState: MapLegendPresentationState
-    let layer: MapLayer
-    let severeItems: [SevereLegendItem]
-    let fireItems: [FireLegendItem]
-    let showsHatchingExplanation: Bool
-
-    var allowsInteraction: Bool {
-        showsHatchingExplanation &&
-        !severeItems.isEmpty &&
-        presentationState != .loading &&
-        presentationState != .confirmedEmpty &&
-        presentationState != .unavailable
-    }
-
-    static func loading(for layer: MapLayer) -> MapLegendState {
-        MapLegendState(
-            presentationState: .loading,
-            layer: layer,
-            severeItems: [],
-            fireItems: [],
-            showsHatchingExplanation: false
-        )
-    }
-
-    static func resolving(
-        for layer: MapLayer,
-        severeItems: [SevereLegendItem],
-        fireItems: [FireLegendItem],
-        showsHatchingExplanation: Bool
-    ) -> MapLegendState {
-        MapLegendState(
-            presentationState: .resolving,
-            layer: layer,
-            severeItems: severeItems,
-            fireItems: fireItems,
-            showsHatchingExplanation: showsHatchingExplanation
-        )
-    }
-
-    static func current(
-        for layer: MapLayer,
-        severeItems: [SevereLegendItem],
-        fireItems: [FireLegendItem],
-        showsHatchingExplanation: Bool
-    ) -> MapLegendState {
-        MapLegendState(
-            presentationState: .current,
-            layer: layer,
-            severeItems: severeItems,
-            fireItems: fireItems,
-            showsHatchingExplanation: showsHatchingExplanation
-        )
-    }
-
-    static func confirmedEmpty(for layer: MapLayer) -> MapLegendState {
-        MapLegendState(
-            presentationState: .confirmedEmpty,
-            layer: layer,
-            severeItems: [],
-            fireItems: [],
-            showsHatchingExplanation: false
-        )
-    }
-
-    static func stale(
-        for layer: MapLayer,
-        severeItems: [SevereLegendItem],
-        fireItems: [FireLegendItem],
-        showsHatchingExplanation: Bool
-    ) -> MapLegendState {
-        MapLegendState(
-            presentationState: .stale,
-            layer: layer,
-            severeItems: severeItems,
-            fireItems: fireItems,
-            showsHatchingExplanation: showsHatchingExplanation
-        )
-    }
-
-    static func unavailable(for layer: MapLayer) -> MapLegendState {
-        MapLegendState(
-            presentationState: .unavailable,
-            layer: layer,
-            severeItems: [],
-            fireItems: [],
-            showsHatchingExplanation: false
-        )
-    }
-
-    func withPresentationState(_ presentationState: MapLegendPresentationState) -> MapLegendState {
-        MapLegendState(
-            presentationState: presentationState,
-            layer: layer,
-            severeItems: severeItems,
-            fireItems: fireItems,
-            showsHatchingExplanation: showsHatchingExplanation
-        )
-    }
-
-    var headlineText: String {
-        switch presentationState {
-        case .loading:
-            "Getting \(layer.legendSubject)…"
-        case .resolving:
-            "\(layer.legendDisplayTitle) · Updating"
-        case .current:
-            layer.legendDisplayTitle
-        case .confirmedEmpty:
-            "No \(layer.legendSubject)"
-        case .stale:
-            "\(layer.legendDisplayTitle) saved locally"
-        case .unavailable:
-            "\(layer.legendDisplayTitle) unavailable"
-        }
-    }
-
-    var voiceOverText: String {
-        switch presentationState {
-        case .loading:
-            "Getting \(layer.legendSubject). Loading."
-        case .resolving:
-            "Updating \(layer.legendSubject). Showing saved data while the refresh completes."
-        case .current:
-            "\(layer.legendDisplayTitle) loaded."
-        case .confirmedEmpty:
-            "No \(layer.legendSubject). Successfully loaded and confirmed empty."
-        case .stale:
-            "\(layer.legendDisplayTitle) saved locally. Refresh failed."
-        case .unavailable:
-            "\(layer.legendDisplayTitle) unavailable. No saved data."
-        }
-    }
-}
-
-enum MapLegendPresentationState: Sendable, Equatable {
-    case loading
-    case resolving
-    case current
-    case confirmedEmpty
-    case stale
-    case unavailable
-}
-
-struct SevereLegendItem: Sendable, Hashable, Identifiable {
-    let id: String
-    let probability: ThreatProbability
-    let fillHex: String?
-    let strokeHex: String?
-}
-
-struct FireLegendItem: Sendable, Hashable, Identifiable {
-    let riskLevel: Int
-    let riskLevelDescription: String
-    let fillHex: String?
-    let strokeHex: String?
-
-    var id: Int { riskLevel }
-}
-
-private struct MapDataPayload: Sendable {
-    let stormRisk: MapFetchOutcome<[StormRiskDTO]>
-    let severeRisks: MapFetchOutcome<[SevereRiskShapeDTO]>
-    let mesos: MapFetchOutcome<[MdDTO]>
-    let fireRisk: MapFetchOutcome<[FireRiskDTO]>
-    let activeWarnings: [ActiveWarningGeometry]
-}
-
-private enum MapFetchOutcome<Value: Sendable>: Sendable {
-    case success(Value)
-    case failure
-    case cancelled
-
-    var isCancellation: Bool {
-        if case .cancelled = self {
-            return true
-        }
-        return false
-    }
-
-    var value: Value? {
-        if case .success(let value) = self {
-            return value
-        }
-        return nil
-    }
-}
-
-struct MapOverlayBuildPlan: Sendable {
-    enum Kind: Sendable {
-        case probability
-        case intensity(level: Int)
-        case warning
-    }
-
-    let key: String
-    let polygonKey: String
-    let kind: Kind
-    let signature: Int
-}
-
-private struct MapLayerRenderPlan: Sendable {
-    let layer: MapLayer
-    let polygonEntries: [MapPolygonEntry]
-    let overlayPlans: [MapOverlayBuildPlan]
-    let overlayRevision: Int
-    let legendState: MapLegendState
-
-    func withLegendState(_ legendState: MapLegendState) -> MapLayerRenderPlan {
-        MapLayerRenderPlan(
-            layer: layer,
-            polygonEntries: polygonEntries,
-            overlayPlans: overlayPlans,
-            overlayRevision: overlayRevision,
-            legendState: legendState
-        )
-    }
-}
-
-private actor MapScenePlanner {
-    func buildRenderPlans(
-        payload: MapDataPayload,
-        existingPlans: [MapLayer: MapLayerRenderPlan],
-        polygonMapper: MapPolygonMapper,
-        warningPolygons: KeyedMapPolygons
-    ) -> [MapLayer: MapLayerRenderPlan] {
-        Dictionary(
-            uniqueKeysWithValues: MapLayer.allCases.map { layer in
-                (
-                    layer,
-                    MapRenderPlanBuilder.build(
-                        layer: layer,
-                        payload: payload,
-                        existingPlan: existingPlans[layer],
-                        polygonMapper: polygonMapper,
-                        warningPolygons: warningPolygons
-                    )
-                )
-            }
-        )
-    }
-}
-
-@MainActor
-private enum MapSceneMaterializer {
-    static func materialize(
-        plan: MapLayerRenderPlan,
-        initialCenterCoordinate: CLLocationCoordinate2D?,
-        showsWarningGeometry: Bool
-    ) -> MapLayerScene {
-        var polygonsByKey: [String: MKPolygon] = [:]
-        polygonsByKey.reserveCapacity(plan.polygonEntries.count)
-
-        for polygonEntry in plan.polygonEntries {
-            polygonsByKey[polygonEntry.key] = polygonEntry.polygon
-        }
-
-        var overlays: [MapOverlayEntry] = []
-        overlays.reserveCapacity(plan.overlayPlans.count)
-
-        for overlayPlan in plan.overlayPlans {
-            if case .warning = overlayPlan.kind, showsWarningGeometry == false {
-                continue
-            }
-
-            guard let polygon = polygonsByKey[overlayPlan.polygonKey] else { continue }
-
-            let overlay: MKOverlay
-            switch overlayPlan.kind {
-            case .probability:
-                overlay = RiskPolygonOverlay.probability(from: polygon, overlayKey: overlayPlan.key)
-            case .intensity(let level):
-                let style = RiskPolygonStyleResolver.probabilityStyle(for: polygon)
-                overlay = RiskPolygonOverlay.intensity(
-                    from: polygon,
-                    level: level,
-                    strokeColor: style.stroke,
-                    fillColor: style.fill,
-                    overlayKey: overlayPlan.key
-                )
-            case .warning:
-                overlay = polygon
-            }
-
-            overlays.append(
-                MapOverlayEntry(
-                    key: overlayPlan.key,
-                    overlay: overlay,
-                    signature: overlayPlan.signature
-                )
-            )
-        }
-
-        return MapLayerScene(
-            canvasState: MapCanvasState(
-                overlays: overlays,
-                overlayRevision: overlayRevision(for: overlays),
-                initialCenterCoordinate: initialCenterCoordinate
-            ),
-            legendState: plan.legendState,
-            warningLegendItems: WarningLegendItem.rendered(from: overlays)
-        )
-    }
-
-    private static func overlayRevision(for overlays: [MapOverlayEntry]) -> Int {
-        var hasher = StableMapHasher()
-        hasher.combine(overlays.count)
-
-        for overlay in overlays {
-            hasher.combine(overlay.key)
-            hasher.combine(overlay.signature)
-        }
-
-        return hasher.intValue
-    }
-}
-
-private enum MapRenderPlanBuilder {
-    static func build(
-        layer: MapLayer,
-        payload: MapDataPayload,
-        existingPlan: MapLayerRenderPlan?,
-        polygonMapper: MapPolygonMapper,
-        warningPolygons: KeyedMapPolygons
-    ) -> MapLayerRenderPlan {
-        switch layer {
-        case .categorical:
-            guard case .success(let stormRisk) = payload.stormRisk else {
-                return buildFailurePlan(
-                    layer: layer,
-                    existingPlan: existingPlan,
-                    warningPolygons: warningPolygons
-                )
-            }
-            let severeRisks = payload.severeRisks.value ?? []
-            let mesos = payload.mesos.value ?? []
-            let fireRisk = payload.fireRisk.value ?? []
-            let mappedPolygons = polygonMapper.polygons(
-                for: layer,
-                stormRisk: stormRisk,
-                severeRisks: severeRisks,
-                mesos: mesos,
-                fires: fireRisk
-            )
-            return buildLoadedPlan(
-                layer: layer,
-                mappedPolygons: mappedPolygons,
-                warningPolygons: warningPolygons,
-                legendState: legendState(
-                    for: layer,
-                    presentationState: stormRisk.isEmpty ? .confirmedEmpty : .current,
-                    severeRisks: severeRisks,
-                    fireRisk: fireRisk
-                )
-            )
-
-        case .wind, .hail, .tornado:
-            guard case .success(let severeRisks) = payload.severeRisks else {
-                return buildFailurePlan(
-                    layer: layer,
-                    existingPlan: existingPlan,
-                    warningPolygons: warningPolygons
-                )
-            }
-            let stormRisk = payload.stormRisk.value ?? []
-            let mesos = payload.mesos.value ?? []
-            let fireRisk = payload.fireRisk.value ?? []
-            let mappedPolygons = polygonMapper.polygons(
-                for: layer,
-                stormRisk: stormRisk,
-                severeRisks: severeRisks,
-                mesos: mesos,
-                fires: fireRisk
-            )
-            return buildLoadedPlan(
-                layer: layer,
-                mappedPolygons: mappedPolygons,
-                warningPolygons: warningPolygons,
-                legendState: legendState(
-                    for: layer,
-                    presentationState: severeRisks.isEmpty ? .confirmedEmpty : .current,
-                    severeRisks: severeRisks,
-                    fireRisk: fireRisk
-                )
-            )
-
-        case .fire:
-            guard case .success(let fireRisk) = payload.fireRisk else {
-                return buildFailurePlan(
-                    layer: layer,
-                    existingPlan: existingPlan,
-                    warningPolygons: warningPolygons
-                )
-            }
-            let stormRisk = payload.stormRisk.value ?? []
-            let severeRisks = payload.severeRisks.value ?? []
-            let mesos = payload.mesos.value ?? []
-            let mappedPolygons = polygonMapper.polygons(
-                for: layer,
-                stormRisk: stormRisk,
-                severeRisks: severeRisks,
-                mesos: mesos,
-                fires: fireRisk
-            )
-            return buildLoadedPlan(
-                layer: layer,
-                mappedPolygons: mappedPolygons,
-                warningPolygons: warningPolygons,
-                legendState: legendState(
-                    for: layer,
-                    presentationState: fireRisk.isEmpty ? .confirmedEmpty : .current,
-                    severeRisks: severeRisks,
-                    fireRisk: fireRisk
-                )
-            )
-
-        case .meso:
-            guard case .success(let mesos) = payload.mesos else {
-                return buildFailurePlan(
-                    layer: layer,
-                    existingPlan: existingPlan,
-                    warningPolygons: warningPolygons
-                )
-            }
-            let stormRisk = payload.stormRisk.value ?? []
-            let severeRisks = payload.severeRisks.value ?? []
-            let fireRisk = payload.fireRisk.value ?? []
-            let mappedPolygons = polygonMapper.polygons(
-                for: layer,
-                stormRisk: stormRisk,
-                severeRisks: severeRisks,
-                mesos: mesos,
-                fires: fireRisk
-            )
-            return buildLoadedPlan(
-                layer: layer,
-                mappedPolygons: mappedPolygons,
-                warningPolygons: warningPolygons,
-                legendState: legendState(
-                    for: layer,
-                    presentationState: mesos.isEmpty ? .confirmedEmpty : .current,
-                    severeRisks: severeRisks,
-                    fireRisk: fireRisk
-                )
-            )
-        }
-    }
-
-    private static func buildLoadedPlan(
-        layer: MapLayer,
-        mappedPolygons: KeyedMapPolygons,
-        warningPolygons: KeyedMapPolygons,
-        legendState: MapLegendState
-    ) -> MapLayerRenderPlan {
-        var probabilityOverlays: [MapOverlayBuildPlan] = []
-        var intensityOverlaysByLevel: [(level: Int, plan: MapOverlayBuildPlan)] = []
-        probabilityOverlays.reserveCapacity(mappedPolygons.keyedPolygons.count)
-
-        for polygonEntry in mappedPolygons.keyedPolygons {
-            let metadata = StormRiskPolygonStyleMetadata.decode(from: polygonEntry.subtitle)
-            if let cigLevel = metadata?.cigLevel {
-                intensityOverlaysByLevel.append(
-                    (
-                        level: cigLevel,
-                        plan: overlayPlan(
-                            for: polygonEntry,
-                            kind: .intensity(level: cigLevel)
-                        )
-                    )
-                )
-            } else {
-                probabilityOverlays.append(
-                    overlayPlan(
-                        for: polygonEntry,
-                        kind: .probability
-                    )
-                )
-            }
-        }
-
-        let orderedIntensityOverlays = intensityOverlaysByLevel
-            .sorted { $0.level < $1.level }
-            .map(\.plan)
-
-        let warningOverlays = warningPolygons.keyedPolygons.map {
-            overlayPlan(
-                for: $0,
-                kind: .warning
-            )
-        }
-        let overlayPlans = probabilityOverlays + orderedIntensityOverlays + warningOverlays
-
-        return MapLayerRenderPlan(
-            layer: layer,
-            polygonEntries: mappedPolygons.keyedPolygons + warningPolygons.keyedPolygons,
-            overlayPlans: overlayPlans,
-            overlayRevision: overlayRevision(for: overlayPlans),
-            legendState: legendState
-        )
-    }
-
-    private static func buildFailurePlan(
-        layer: MapLayer,
-        existingPlan: MapLayerRenderPlan?,
-        warningPolygons: KeyedMapPolygons
-    ) -> MapLayerRenderPlan {
-        guard let existingPlan else {
-            return warningPolygons.keyedPolygons.isEmpty
-                ? MapLayerRenderPlan(
-                    layer: layer,
-                    polygonEntries: [],
-                    overlayPlans: [],
-                    overlayRevision: 0,
-                    legendState: .unavailable(for: layer)
-                )
-                : buildWarningOnlyFailurePlan(
-                    layer: layer,
-                    warningPolygons: warningPolygons
-                )
-        }
-
-        if warningPolygons.keyedPolygons.isEmpty == false {
-            return buildWarningPreservingFailurePlan(
-                layer: layer,
-                existingPlan: existingPlan,
-                warningPolygons: warningPolygons
-            )
-        }
-
-        switch existingPlan.legendState.presentationState {
-        case .current, .resolving, .stale:
-            return existingPlan.withLegendState(
-                existingPlan.legendState.withPresentationState(.stale)
-            )
-        case .loading, .confirmedEmpty, .unavailable:
-            return MapLayerRenderPlan(
-                layer: layer,
-                polygonEntries: [],
-                overlayPlans: [],
-                overlayRevision: 0,
-                legendState: .unavailable(for: layer)
-            )
-        }
-    }
-
-    private static func buildWarningOnlyFailurePlan(
-        layer: MapLayer,
-        warningPolygons: KeyedMapPolygons
-    ) -> MapLayerRenderPlan {
-        let warningOverlayPlans = warningPolygons.keyedPolygons.map {
-            overlayPlan(for: $0, kind: .warning)
-        }
-
-        return MapLayerRenderPlan(
-            layer: layer,
-            polygonEntries: warningPolygons.keyedPolygons,
-            overlayPlans: warningOverlayPlans,
-            overlayRevision: overlayRevision(for: warningOverlayPlans),
-            legendState: .unavailable(for: layer)
-        )
-    }
-
-    private static func buildWarningPreservingFailurePlan(
-        layer: MapLayer,
-        existingPlan: MapLayerRenderPlan,
-        warningPolygons: KeyedMapPolygons
-    ) -> MapLayerRenderPlan {
-        let existingWarningPolygonKeys = Set(
-            existingPlan.overlayPlans.compactMap {
-                if case .warning = $0.kind {
-                    $0.polygonKey
-                } else {
-                    nil
-                }
-            }
-        )
-
-        let preservedPolygons = existingPlan.polygonEntries.filter {
-            existingWarningPolygonKeys.contains($0.key) == false
-        }
-        let warningOverlayPlans = warningPolygons.keyedPolygons.map {
-            overlayPlan(for: $0, kind: .warning)
-        }
-        let preservedOverlayPlans = existingPlan.overlayPlans.filter {
-            if case .warning = $0.kind {
-                return false
-            }
-            return true
-        }
-
-        let legendState: MapLegendState
-        switch existingPlan.legendState.presentationState {
-        case .current, .resolving, .stale:
-            legendState = existingPlan.legendState.withPresentationState(.stale)
-        case .loading, .confirmedEmpty, .unavailable:
-            legendState = .unavailable(for: layer)
-        }
-
-        let overlayPlans = preservedOverlayPlans + warningOverlayPlans
-        return MapLayerRenderPlan(
-            layer: layer,
-            polygonEntries: preservedPolygons + warningPolygons.keyedPolygons,
-            overlayPlans: overlayPlans,
-            overlayRevision: overlayRevision(for: overlayPlans),
-            legendState: legendState
-        )
-    }
-
-    private static func overlayPlan(
-        for polygonEntry: MapPolygonEntry,
-        kind: MapOverlayBuildPlan.Kind
-    ) -> MapOverlayBuildPlan {
-        let key: String
-        switch kind {
-        case .probability:
-            key = "\(polygonEntry.key)|probability"
-        case .intensity(let level):
-            key = "\(polygonEntry.key)|intensity|\(level)"
-        case .warning:
-            key = "\(polygonEntry.key)|warning"
-        }
-
-        return MapOverlayBuildPlan(
-            key: key,
-            polygonKey: polygonEntry.key,
-            kind: kind,
-            signature: overlaySignature(
-                key: key,
-                subtitle: polygonEntry.subtitle,
-                kind: kind
-            )
-        )
-    }
-
-    private static func overlaySignature(
-        key: String,
-        subtitle: String?,
-        kind: MapOverlayBuildPlan.Kind
-    ) -> Int {
-        var hasher = StableMapHasher()
-        hasher.combine(key)
-        hasher.combine(subtitle)
-
-        switch kind {
-        case .probability:
-            hasher.combine("probability")
-        case .intensity(let level):
-            hasher.combine("intensity")
-            hasher.combine(level)
-        case .warning:
-            hasher.combine("warning")
-        }
-
-        return hasher.intValue
-    }
-
-    private static func overlayRevision(for plans: [MapOverlayBuildPlan]) -> Int {
-        var hasher = StableMapHasher()
-        hasher.combine(plans.count)
-        for plan in plans {
-            hasher.combine(plan.key)
-            hasher.combine(plan.signature)
-        }
-        return hasher.intValue
-    }
-
-    private static func legendState(
-        for layer: MapLayer,
-        presentationState: MapLegendPresentationState,
-        severeRisks: [SevereRiskShapeDTO],
-        fireRisk: [FireRiskDTO]
-    ) -> MapLegendState {
-        switch layer {
-        case .fire:
-            let items = fireLegendItems(from: fireRisk)
-            if presentationState == .confirmedEmpty {
-                return .confirmedEmpty(for: layer)
-            }
-
-            return MapLegendState(
-                presentationState: presentationState,
-                layer: layer,
-                severeItems: [],
-                fireItems: items,
-                showsHatchingExplanation: false
-            )
-
-        case .tornado, .hail, .wind:
-            let severeItems = severeLegendItems(for: layer, severeRisks: severeRisks)
-            let hasHatching = severeRisks
-                .filter { $0.type == layer.threatType }
-                .contains { $0.intensityLevel != nil }
-
-            if presentationState == .confirmedEmpty {
-                return .confirmedEmpty(for: layer)
-            }
-
-            return MapLegendState(
-                presentationState: presentationState,
-                layer: layer,
-                severeItems: severeItems,
-                fireItems: [],
-                showsHatchingExplanation: hasHatching
-            )
-
-        default:
-            if presentationState == .confirmedEmpty {
-                return .confirmedEmpty(for: layer)
-            }
-
-            return MapLegendState(
-                presentationState: presentationState,
-                layer: layer,
-                severeItems: [],
-                fireItems: [],
-                showsHatchingExplanation: false
-            )
-        }
-    }
-
-    private static func fireLegendItems(from fireRisk: [FireRiskDTO]) -> [FireLegendItem] {
-        let mostRecentByLevel = Dictionary(
-            fireRisk.map {
-                (
-                    $0.riskLevel,
-                    FireLegendItem(
-                        riskLevel: $0.riskLevel,
-                        riskLevelDescription: $0.riskLevelDescription,
-                        fillHex: $0.fill,
-                        strokeHex: $0.stroke
-                    )
-                )
-            },
-            uniquingKeysWith: { lhs, _ in lhs }
-        )
-
-        return mostRecentByLevel.values.sorted { $0.riskLevel > $1.riskLevel }
-    }
-
-    private static func severeLegendItems(
-        for layer: MapLayer,
-        severeRisks: [SevereRiskShapeDTO]
-    ) -> [SevereLegendItem] {
-        let filtered = severeRisks.filter { risk in
-            guard risk.type == layer.threatType else { return false }
-            guard risk.intensityLevel == nil else { return false }
-
-            if case .percent(let value) = risk.probabilities, value <= 0 {
-                return false
-            }
-
-            return true
-        }
-
-        let dedupedByTitle = Dictionary(
-            filtered.map {
-                (
-                    $0.title,
-                    SevereLegendItem(
-                        id: $0.title,
-                        probability: $0.probabilities,
-                        fillHex: $0.fill,
-                        strokeHex: $0.stroke
-                    )
-                )
-            },
-            uniquingKeysWith: { lhs, _ in lhs }
-        )
-
-        return dedupedByTitle.values.sorted {
-            let lhsProbability = $0.probability.intValue
-            let rhsProbability = $1.probability.intValue
-            if lhsProbability != rhsProbability {
-                return lhsProbability < rhsProbability
-            }
-
-            let lhsSignificanceRank = isSignificant($0.probability) ? 1 : 0
-            let rhsSignificanceRank = isSignificant($1.probability) ? 1 : 0
-            if lhsSignificanceRank != rhsSignificanceRank {
-                return lhsSignificanceRank < rhsSignificanceRank
-            }
-
-            return $0.id < $1.id
-        }
-    }
-
-    private static func isSignificant(_ probability: ThreatProbability) -> Bool {
-        if case .significant = probability {
-            return true
-        }
-        return false
-    }
-}
-
-private extension Result {
-    var isCancellation: Bool {
-        if case .failure(let error) = self, error is CancellationError {
-            return true
-        }
-        return false
-    }
-}
-
-private extension Result where Success == [SevereRiskShapeDTO], Failure == Error {
-    func value(orLogging label: StaticString, logger: Logger) -> [SevereRiskShapeDTO] {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure(let error):
-            if error is CancellationError { return [] }
-            logger.error("Failed to load \(label): \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
-}
-
-private extension Result where Success == [StormRiskDTO], Failure == Error {
-    func value(orLogging label: StaticString, logger: Logger) -> [StormRiskDTO] {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure(let error):
-            if error is CancellationError { return [] }
-            logger.error("Failed to load \(label): \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
-}
-
-private extension Result where Success == [MdDTO], Failure == Error {
-    func value(orLogging label: StaticString, logger: Logger) -> [MdDTO] {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure(let error):
-            if error is CancellationError { return [] }
-            logger.error("Failed to load \(label): \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
-}
-
-private extension Result where Success == [FireRiskDTO], Failure == Error {
-    func value(orLogging label: StaticString, logger: Logger) -> [FireRiskDTO] {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure(let error):
-            if error is CancellationError { return [] }
-            logger.error("Failed to load \(label): \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
-}
-
-private extension MapLayer {
-    var legendDisplayTitle: String {
-        switch self {
-        case .categorical:
-            return "Severe Risk"
-        case .wind:
-            return "Wind Risk"
-        case .hail:
-            return "Hail Risk"
-        case .tornado:
-            return "Tornado Risk"
-        case .meso:
-            return "Mesoscale"
-        case .fire:
-            return "Fire Risk"
-        }
-    }
-
-    var legendSubject: String {
-        switch self {
-        case .categorical:
-            return "severe risk"
-        case .wind:
-            return "wind risk"
-        case .hail:
-            return "hail risk"
-        case .tornado:
-            return "tornado risk"
-        case .meso:
-            return "mesoscale"
-        case .fire:
-            return "fire risk"
-        }
-    }
-
-    var threatType: ThreatType {
-        switch self {
-        case .tornado:
-            return .tornado
-        case .hail:
-            return .hail
-        case .wind:
-            return .wind
-        default:
-            return .unknown
-        }
-    }
-}
